PEP8 style check.

+++ test/lenet.py
@@ -155,7 +155,7 @@ def lenet(dataset, learning_rate, epochs, batch_size, n_hidden, kernels):
 
             if (iteration + 1) % valid_freq == 0:
                 valid_losses = [valid_model(i) for i in range(n_batches_valid)]
-                current_valid_loss = np.mean(valid_losses)
+                current_valid_loss = float(np.mean(valid_losses))
 
                 print(
                     'epoch %i, batch %i/%i, batch average cost %f, validation error %f %%' %
